+++ projects/plugins/jetpack/tests/e2e/specs/post-connection/waf-blocking.test.js
@@ -1,19 +1,15 @@
-import { prerequisitesBuilder } from '_jetpack-e2e-commons/env/index.js';
 import { test, expect } from '_jetpack-e2e-commons/fixtures/base-test.ts';
 import { WpPage } from '_jetpack-e2e-commons/pages/index.js';
 import { enableAutomaticRules, generateRules } from '../../helpers/waf-helper.js';
-import playwrightConfig from '../../playwright.config.mjs';
 
 test.describe.parallel( 'WAF Blocking', () => {
-	test.beforeAll( async ( { browser } ) => {
-		const page = await browser.newPage( playwrightConfig.use );
+	test.beforeAll( async ( { testUtils } ) => {
 		/* Note that having the Complete plan does not really apply yet because we are mocking the data returned from
 		 * the API for now. See tools/e2e-commons/plugins/e2e-waf-data-interceptor.php for details.
 		 */
-		await prerequisitesBuilder( page ).withActiveModules( [ 'waf' ] ).build();
+		await testUtils.activateModule( 'waf' );
 		await enableAutomaticRules();
 		await generateRules();
-		await page.close();
 	} );
 
 	test( 'Block a simple request', async ( { page, baseURL } ) => {