+++ projects/plugins/social/tests/e2e/specs/admin-page.test.js
@@ -1,10 +1,9 @@
 import { prerequisitesBuilder } from '_jetpack-e2e-commons/env/prerequisites.js';
 import { expect, test } from '_jetpack-e2e-commons/fixtures/base-test.ts';
 import logger from '_jetpack-e2e-commons/logger.js';
-import { disconnect } from '_jetpack-e2e-commons/utils/index.ts';
 
-test.beforeEach( async ( { page, requestUtils } ) => {
-	await disconnect( requestUtils );
+test.beforeEach( async ( { page, testUtils } ) => {
+	await testUtils.disconnect();
 	await prerequisitesBuilder( page ).withActivePlugins( [ 'jetpack-social' ] ).build();
 } );